isomakerimport: More rock ridge code.
svn:sync-xref-src-repo-rev: r116709

+++ include/iprt/formats/iso9660.h
@@ -1,4 +1,4 @@
-/* $Id: iso9660.h 67788 2017-07-05 10:43:52Z knut.osmundsen@oracle.com $ */
+/* $Id: iso9660.h 67790 2017-07-05 11:24:53Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT, ISO 9660 File System
  */
@@ -1158,15 +1158,15 @@ typedef ISO9660RRIPTF const *PCISO9660RRIPTF;
  */
 DECLINLINE(uint8_t) Iso9660RripTfCalcLength(uint8_t fFlags)
 {
-    uint8_t cTimestamps = ((fFlags & ISO9660RRIPTF_F_BIRTH)      != 0)
-                        + ((fFlags & ISO9660RRIPTF_F_MODIFY)     != 0)
-                        + ((fFlags & ISO9660RRIPTF_F_ACCESS)     != 0)
-                        + ((fFlags & ISO9660RRIPTF_F_CHANGE)     != 0)
-                        + ((fFlags & ISO9660RRIPTF_F_BACKUP)     != 0)
-                        + ((fFlags & ISO9660RRIPTF_F_EXPIRATION) != 0)
-                        + ((fFlags & ISO9660RRIPTF_F_EFFECTIVE)  != 0);
-    return cTimestamps * (uint8_t)(fFlags & ISO9660RRIPTF_F_LONG_FORM ? sizeof(ISO9660TIMESTAMP) : sizeof(ISO9660RECTIMESTAMP))
-         + (uint8_t)RT_OFFSETOF(ISO9660RRIPTF, abPayload);
+    unsigned cTimestamps = ((fFlags & ISO9660RRIPTF_F_BIRTH)      != 0)
+                         + ((fFlags & ISO9660RRIPTF_F_MODIFY)     != 0)
+                         + ((fFlags & ISO9660RRIPTF_F_ACCESS)     != 0)
+                         + ((fFlags & ISO9660RRIPTF_F_CHANGE)     != 0)
+                         + ((fFlags & ISO9660RRIPTF_F_BACKUP)     != 0)
+                         + ((fFlags & ISO9660RRIPTF_F_EXPIRATION) != 0)
+                         + ((fFlags & ISO9660RRIPTF_F_EFFECTIVE)  != 0);
+    return (uint8_t)(  cTimestamps * (fFlags & ISO9660RRIPTF_F_LONG_FORM ? sizeof(ISO9660TIMESTAMP) : sizeof(ISO9660RECTIMESTAMP))
+                     + RT_OFFSETOF(ISO9660RRIPTF, abPayload));
 }
